+++ lib/src/controls/inputs/buttons/outlined_button.dart
@@ -17,6 +17,7 @@ class OutlinedButton extends BaseButton {
     FocusNode? focusNode,
     bool autofocus = false,
     ButtonStyle? style,
+    bool focusable = true,
   }) : super(
           key: key,
           child: child,
@@ -25,6 +26,7 @@ class OutlinedButton extends BaseButton {
           onLongPress: onLongPress,
           onPressed: onPressed,
           style: style,
+          focusable: focusable,
         );
 
   @override